requirements.txt
boto3==1.9.189
botocore==1.12.189
docutils==0.14
jmespath==0.9.4
joblib==0.13.2
numpy==1.16.4
pandas==0.24.2
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
urllib3==1.25.3